app/requirements.txt
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Compress==1.8.0
Flask-Cors==3.0.10
requests==2.25.1
moment
boto3
simplejson
pyyaml
marko
pymysql